hadoop 版本

+++ recommend-engine/src/main/java/com/babyduncan/TestDataGenerator.java
@@ -32,7 +32,7 @@ public class TestDataGenerator {
 
     }
 
-    private static float generateScore() {
+    private static int generateScore() {
         return new Random().nextInt(10);
     }
